Report E_INVALID_OPERATOR with the right order of operands
E.g.,

  "Invalid operator: ++ string."

or

  "Invalid operator: ! string."

As a hack, POSTINC_OP and POSTDEC_OP (for post-increment and post-decrement) have been introduced as tokens, even if not used in the grammar, so they can be passed as an operator to the AST.

This allows us to distinguish prefix and postfix ++ and --, to show e.g.

  "Invalid operator: string --."

when appropriate.